fix nasty bug around scala implicits
the current code generator does not generate implicits for writing date
times that work. The compiler does not apply the implicits when they
are private, but it will apply them if they are package private to
the toplevel generated package.

NOTE: making them private to models or json does not fix!

- remove dangerous imports of everything in the play json package.
  just import what is needed

- do not import from the syntax package multiple times