Updating swagger parser to 1.0.24
- The vast majority of the changes is around changing the
swagger imports since the classes moved to a new org/package
structure with the new release
- We seemed to have a missing mapping for when integers were
defined without a format on a swagger spec. I grabbed the chance
to fix that by changing SchemaType to explicitly have fields
for the swagger type and the format. This means the mapping to an
apidoc type is more explicit
- Also we were throwing an error deliberately when a composite model
had interfaces, I'm thinking that because we were assuming the previous
parser didn't support those? The new one supports, and seems like our
Parser is already covering that case (Parser.scala:84)