Handle gracefully Swagger spec with no model types in 200 response.
Currently Swagger import fails (with not exhaustive Scala pattern matching error) if at least one of the paths' 200 response has no model type (example: string or array of string response body).
Cases like this are now ignored by ApiBuilder when looking for Resources: no errors are thrown but that resource and related paths/operations are not be imported in ApiBuilder (they cannot be imported being not objects, but at least the rest of the Swagger spec will be imported).