Introduce inline types
  - This change allows user to declare types inline when using api.json format
  - Wherever we declare a type as the name of the model - we can now instead
    declare the type as an object, defining the enum, model or union inline
  - Tests have examples of varying inlining

  - Consider this feature experimental

As one example, this change allows you to define a model
that contains a single field named 'status' that is defined
as an enum named 'user_status'. Instead of having to declare
the enum itself, you can inline the enum into the definition
of the field:
```
    {
      "name": "Api Doc",

      "models": {
        "user": {
          "fields": [
            {
             "name": "status",
              "type": {
                "enum": "user_status",
                "values": [
                  { "name": "guest" },
                  { "name": "registered" }
                ]
              }
            }
          ]
        }
      }
    }
```

API Builder will automatically expand the type, creating the enum - ie. this feature
is to support human editing and does not change in any way the underling service specification.